cook to render targets
